flast-backend/requirements.txt
flask-security-too
sqlalchemy
flask
bcrypt
flask_wtf
flask_cors
flask_mail
bleach